Move CLI report models into CLISupport
Move Codable CLI status/install/inspect DTOs out of KeyPathAppKit and into KeyPathCLISupport so their JSON contract can be exercised without compiling the app-heavy target.

Keep the AppKit-specific report builders as extensions in SystemFacade and re-export the support module for existing KeyPathAppKit consumers.

Validation: KEYPATH_SNAPSHOTS=0 ./Scripts/run-tests-safe.sh; TEST_FILTER='CLIReportModelsTests' ./Scripts/run-tests-safe.sh; TEST_FILTER='CLIOutputContractTests|CLIOutputSnapshotTests' ./Scripts/run-tests-safe.sh; swift build --target KeyPathCLISupport --target KeyPathAppKit --target KeyPathCLI; swift build --target KeyPathCLISupport; swiftformat --lint changed files; git diff --check; python3 Scripts/check-accessibility.py. Review gate: ./Scripts/review-gate.sh selected remote review gate.
